Add files via upload
I've managed to do something which allows the user to create a new file in the directory, I think it's a bit convoluted and could be tidied up somehow but it works for now. I also found out that we don't need to ask the user for the file extension because the script will automatically assume it's a txt file when you use "touch" to create a new file. I created two functions and called them in the menu to tidy up the code